requirements.txt
Django==2.0.4
pandas==0.24.2
dill==0.2.9
djangorestframework==3.9.2
jupyterlab==0.35.4
requests==2.21.0
sklearn==0.0
